Convert 'int i' to 'size_t i' in src/parallels/ files
Convert the type of loop iterators named 'i', 'j', k',
'ii', 'jj', 'kk', to be 'size_t' instead of 'int' or
'unsigned int', also santizing 'ii', 'jj', 'kk' to use
the normal 'i', 'j', 'k' naming

+++ src/parallels/parallels_storage.c
@@ -129,14 +129,14 @@ static char *parallelsMakePoolName(virConnectPtr conn, const char *path)
 {
     parallelsConnPtr privconn = conn->privateData;
     char *name;
-    unsigned int i;
+    size_t i;
 
     for (i = 0; i < UINT_MAX; i++) {
         bool found = false;
-        int j;
+        size_t j;
 
         if ((!i && VIR_STRDUP(name, path) < 0) ||
-            (i && virAsprintf(&name, "%s-%u", path, i) < 0))
+            (i && virAsprintf(&name, "%s-%zu", path, i) < 0))
             return NULL;
 
         for (j = 0; j < strlen(name); j++)
@@ -216,7 +216,7 @@ parallelsPoolAddByDomain(virConnectPtr conn, virDomainObjPtr dom)
     virStoragePoolObjListPtr pools = &privconn->pools;
     char *poolPath;
     virStoragePoolObjPtr pool = NULL;
-    int j;
+    size_t j;
 
     poolPath = mdir_name(pdom->home);
     if (!poolPath) {
